changed page name to use dump, bug fixes for other datasources

+++ data-pipeline/ingestion/wikimedia_dumps/dump_manager.py
@@ -14,6 +14,9 @@ class DumpManager:
         
     def download_page_view_dump(self, date: Date) -> Future[str]:
         def download_page(date: Date) -> str:
+            #check if file exists
+            if os.path.exists(f"dumps/pageviews-{date.year}{date.month:02}{date.day:02}-user"):
+                return f"dumps/pageviews-{date.year}{date.month:02}{date.day:02}-user"
             print("downloading page view dump for date", date)
             response = requests.get(f"https://dumps.wikimedia.org/other/pageview_complete/{date.year}/{date.year}-{date.month:02}/pageviews-{date.year}{date.month:02}{date.day:02}-user.bz2")
             os.makedirs("dumps", exist_ok=True)
@@ -38,8 +41,12 @@ class DumpManager:
 
 
     def get_page_revision_dump_filename(self, date: Date) -> str:
-        # https://dumps.wikimedia.org/other/mediawiki_history/2024-10/enwiki/2024-10.enwiki.2024-11.tsv.bz2
+        #check if file exists
+        if os.path.exists(f"dumps/2024-10.enwiki.{date.year}-{date.month:02}.tsv"):
+            return f"dumps/2024-10.enwiki.{date.year}-{date.month:02}.tsv"
+
         print("downloading page revision dump for date", date)
+        # https://dumps.wikimedia.org/other/mediawiki_history/2024-10/enwiki/2024-10.enwiki.2024-11.tsv.bz2
         response = requests.get(f"https://dumps.wikimedia.org/other/mediawiki_history/2024-10/enwiki/2024-10.enwiki.{date.year}-{date.month:02}.tsv.bz2")
         os.makedirs("dumps", exist_ok=True)
         with open(f"dumps/2024-10.enwiki.{date.year}-{date.month:02}.tsv.bz2", "wb") as file:
